requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
waitress==3.0.2 
gunicorn==20.1.0
earthengine-api==0.1.384
google-cloud-storage==2.10.0
pandas==2.1.3
sqlalchemy==2.0.23
apscheduler==3.10.4
python-dateutil==2.8.2
geopandas==0.14.1
shapely==2.0.2
tensorflow==2.20.0
Pillow==11.2.1